Add an extra secure mode
This automatically wraps outputs coming out of sf functions, and makes inputs to them immutable.

This should make it significantly harder for an addon author to screw something up, although it's still possible. This could be sped up dramatically if we used the documentation we have on functions, to avoid usage of unpack() and to potentially use the specific unwraps instead of object unwrapping.

Could also inline the use of instance.Sanitize.