Progress
Immutability doesn't entirely work so I scrapped it. Need to think of a way to do it while still being able to unwrap it properly. Might need to modify the unwrap function.

Now the wrappers are specialized and cover 95% of all the functions, the rest use a generic vararg solution as before. The slowdown was ~3x (even if broken). Pretty good

Another issue is that this will mess with stack levels.