Add examples to the NSF Helper, modularize tree
Now every view exports a function for the item they want to contribute to the tree on the left panel which decouples things nicely.

This implements view of examples (also viewable in game, but now online too.) highlighted with highlight.js which should hopefully work in game since it is an old library itself.